+++ crates/common/src/config.rs
@@ -2,5 +2,5 @@
 #[serde(default, deny_unknown_fields, rename_all = "kebab-case")]
 pub struct PluginConfig {
     pub bundled_npm: bool,
-    pub intercept_globals: bool,
+    pub shared_globals_dir: bool,
 }